refactor(loader): extract qdrant helpers into module
